requirements.txt
transformers
torch
pytesseract
torchvision
opencv-python
pillow
streamlit